Full progress checkpoint Been making a lot of changes, thought I'd commit them Name change from RB3-X360-Keyboard to RB3_X360_Keyboard Move ViGEmBus check back to MainWindow.cs Fill out most of MainWindow.cs with the main logic loop Get a majority of the main window done Create a pedal press texture (not in the window yet) Lots of changes to XInputPolling.cs: - apparently i didn't commit XInputPolling.cs when i moved InputState to its own file - InitializePoll returns the initialized controller - Poll takes an input controller and returns the current and previous allocated state InputState.cs: changed range array to use bytes instead of ints XInputOut.cs: - Initialize returns the output controller - Output takes an output controller and the current state - Panic takes the output controller from Output - Moved drum mode stuff to a separate function DrumModeOutput KeyboardOut.cs: - Output takes a current and previous state - Add call to panic function - Separate Start and Back into their own checks, instead of doing a combined check - Same for d-pad MIDIOut.cs: - Remove variables such as octave and drum mode, these are handled in and sent from the main window - Output takes current state, previous state, MIDI device reference, and pedal mode, octave, program, and drum mode settings - Move Output above OctaveSwitch - Rename OctaveSwitch to SetOctave - Initialization moved to MainWindow.cs - Panic function takes the output device reference from Output - SetOctave changed from switch-case to procedure, minus drum mode numbers X360 Keys Inputs.txt: - Expanded LT and RT into Left Trigger and Right Trigger - Put underscores instead of spaces as separators for the binary representations - Evened out spacing - A couple small corrections Other small things I left out/may have missed

+++ InOut/InputState.cs
@@ -8,7 +8,7 @@ namespace InOut
 		/// <summary>
 		/// Array of the four ranges the keyboard splits its key bitmask into.
 		/// </summary>
-		public int[] range = new int[4];
+		public byte[] range = new byte[4];
 
 		/// <summary>
 		/// Array of the keyboard's bitmask.
